requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
pytest==7.3.1
selenium==4.28.1
requests==2.32.3
